requirements.txt
# 1. django version >= 5.0.1, < 6.0.0
django>=5.0.1,<6.0.0
# 2. django rest framework version >= 3.14.0, < 4.0.0
djangorestframework>=3.14.0,<4.0.0
# 3. psycopg2( postgresql adapter) version >= 2.9.9, < 3.0.0
psycopg2>=2.9.9,<3.0.0
# 4. drf-spectacular version >= 0.27.1, < 0.28.0
drf-spectacular>=0.27.1,<0.28.0
# 5. channels version >=4.0.0, <5.0.0
channels>=4.0.0,<5.0.0